+++ src/main/kotlin/com/pravo/pravo/global/error/exception/BaseException.java
@@ -2,6 +2,8 @@ package com.pravo.pravo.global.error.exception;
 
 import com.pravo.pravo.global.error.ErrorCode;
 import com.pravo.pravo.global.error.ErrorCode;
+import com.pravo.pravo.global.error.ErrorCode;
+import com.pravo.pravo.global.error.ErrorCode;
 
 public class BaseException extends RuntimeException {
     private final ErrorCode errorCode;
